WIP merge enter and check name steps pt2
