+++ ui/Modal/Org/ConfigureEns/EnterEnsName.svelte
@@ -5,31 +5,39 @@
  with Radicle Linking Exception. For full terms see the included
  LICENSE file.
 -->
-<script lang="typescript" context="module">
-  export interface Result {
-    registration?: ensResolver.Registration;
-    name: string;
-  }
-</script>
-
 <script lang="typescript">
-  import type * as ethers from "ethers";
+  import * as ethers from "ethers";
+
+  import { unreachable } from "ui/src/unreachable";
   import * as svelteStore from "ui/src/svelteStore";
   import * as wallet from "ui/src/wallet";
 
   import { Modal, TextInput } from "ui/DesignSystem";
 
+  import ConfirmRegistration from "./ConfirmRegistration.svelte";
   import ButtonRow from "./shared/ButtonRow.svelte";
 
   import * as ensRegistrar from "ui/src/org/ensRegistrar";
   import * as ensResolver from "ui/src/org/ensResolver";
+  import * as error from "ui/src/error";
   import * as validation from "ui/src/validation";
 
-  export let onSubmit: (result: Result) => void;
-  export let currentName: string | undefined;
+  export let done: () => void;
+  export let name: string = "";
   export let fee: ethers.BigNumber;
 
-  let name = currentName || "";
+  type State =
+    | {
+        type: "commit";
+      }
+    | {
+        type: "register";
+        commitmentSalt: Uint8Array;
+        commitmentBlock: number;
+        minAge: number;
+      };
+
+  let state: State = { type: "commit" };
 
   let validationStatus: validation.ValidationState = {
     status: validation.ValidationStatus.NotStarted,
@@ -37,6 +45,7 @@
 
   let timeoutHandle: NodeJS.Timeout;
   let userInputStarted: boolean = name !== "";
+  let nextStep: "registerName" | "updateMetadata";
 
   function validateName(name: string | undefined): void {
     if (!userInputStarted) {
@@ -79,8 +88,8 @@
         return;
       }
 
-      // Show the green checkmark on the right side of the input box.
       validationStatus = { status: validation.ValidationStatus.Success };
+      nextStep = "registerName";
     } else {
       const registration = await ensResolver.getRegistration(
         `${name}.${ensResolver.DOMAIN}`
@@ -90,6 +99,7 @@
 
       if (registration && registration.owner === walletStore.getAddress()) {
         validationStatus = { status: validation.ValidationStatus.Success };
+        nextStep = "updateMetadata";
         return;
       }
 
@@ -101,40 +111,74 @@
   }
 
   async function handleSubmit(): Promise<void> {
-    /*   const walletStore = svelteStore.get(wallet.store); */
-    /*   const registration = await ensResolver.getRegistration( */
-    /*     `${name}.${ensResolver.DOMAIN}` */
-    /*   ); */
-    /*   if (registration && registration.owner === walletStore.getAddress()) { */
-    /*     onSubmit({ */
-    /*       registration, */
-    /*       name, */
-    /*     }); */
-    /*   } else { */
-    /*     onSubmit({ */
-    /*       name, */
-    /*     }); */
-    /*   } */
+    if (nextStep === "registerName") {
+      commit();
+    } else {
+      done();
+    }
+  }
+
+  async function commit() {
+    // TODO: show notification to confirm tx in wallet
+
+    try {
+      const salt = ethers.utils.randomBytes(32);
+
+      const walletStore = svelteStore.get(wallet.store);
+
+      const commitResult = await ensRegistrar.commit(
+        walletStore.environment,
+        name,
+        salt,
+        fee
+      );
+
+      state = {
+        type: "register",
+        minAge: commitResult.minAge,
+        commitmentBlock: commitResult.receipt.blockNumber,
+        commitmentSalt: salt,
+      };
+    } catch (err) {
+      error.show(
+        new error.Error({
+          message:
+            "Transaction failed. Please try again and confirm the signature & transaction in your connected wallet.",
+          source: err,
+        })
+      );
+    }
   }
 
   $: validateName(name);
 </script>
 
-<Modal
-  emoji="📇"
-  title="Let’s name your org"
-  desc="What should your org be called? This name will show up on the top of your profile and anywhere you interact as an org on Radicle.">
-  <TextInput
-    bind:value={name}
-    showSuccessCheck
-    validation={validationStatus}
-    suffix={`.${ensResolver.DOMAIN}`}
-    placeholder="Your organization name"
-    style="margin: 16px auto; width: 352px;" />
+{#if state.type === "commit"}
+  <Modal
+    emoji="📇"
+    title="Let’s name your org"
+    desc="What should your org be called? This name will show up on the top of your profile and anywhere you interact as an org on Radicle.">
+    <TextInput
+      bind:value={name}
+      showSuccessCheck
+      validation={validationStatus}
+      suffix={`.${ensResolver.DOMAIN}`}
+      placeholder="Your organization name"
+      style="margin: 16px auto; width: 352px;" />
 
-  <ButtonRow
-    onSubmit={handleSubmit}
-    confirmCopy="Continue"
-    disableButtons={validationStatus.status !==
-      validation.ValidationStatus.Success} />
-</Modal>
+    <ButtonRow
+      onSubmit={handleSubmit}
+      confirmCopy="Continue"
+      disableButtons={validationStatus.status !==
+        validation.ValidationStatus.Success} />
+  </Modal>
+{:else if state.type === "register"}
+  <ConfirmRegistration
+    {name}
+    commitmentSalt={state.commitmentSalt}
+    commitmentBlock={state.commitmentBlock}
+    minAge={state.minAge}
+    {done} />
+{:else}
+  {unreachable(state)}
+{/if}
